Added a bunch of stuff
STILL NEED TO COMMENT CODE:
Multiview is *kind of* working. everything is still manual in the editor.
In order to show and scale correctly you need give a source the "ActiveNdi" tag and drag it into the Panel object in the canvas. I built out a bunch of the update unstions in the NDIManger.cs and it attaches to , and get ready for this, the NDI Manager. *shocked pickachu face*

anyway, Good Shit.